// apps/hms-api/src/@generated/prisma-nestjs-graphql/location/location-count-order-by-aggregate.input.ts
import { Field, InputType } from '@nestjs/graphql';
import { SortOrder } from '../prisma/sort-order.enum';

@InputType()
export class LocationCountOrderByAggregateInput {
  @Field(() => SortOrder, { nullable: true })
  id?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  deltek_id?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  name?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  postal_code?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  lat?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  lon?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  region?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  created_at?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  updated_at?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  deleted_at?: keyof typeof SortOrder;
}